Fixed plugin delete issue.

+++ uninstall.php
@@ -13,18 +13,32 @@ if (! defined('WP_UNINSTALL_PLUGIN')) {
 }
 
 /**
- * Delete all pages and settings when plugin in uninstalled.
+ * Delete all pages and settings when plugin is uninstalled.
  */
 function wpfep_delete_options()
 {
     $wpfep_uninstall = get_option('wpfep_general');
-    if ('on' == $wpfep_uninstall['wpfep_remove_data_on_uninstall']) {
+
+    // Check if $wpfep_uninstall is an array
+    if (is_array($wpfep_uninstall) && 'on' == $wpfep_uninstall['wpfep_remove_data_on_uninstall']) {
         // Delete Pages.
         $wpfep_options = get_option('wpfep_profile');
-        wp_delete_post($wpfep_options['login_page'], true);
-        wp_delete_post($wpfep_options['register_page'], true);
-        wp_delete_post($wpfep_options['edit_page'], true);
-        wp_delete_post($wpfep_options['profile_page'], true);
+
+        // Check if $wpfep_options is an array before accessing its indexes
+        if (is_array($wpfep_options)) {
+            if (isset($wpfep_options['login_page'])) {
+                wp_delete_post($wpfep_options['login_page'], true);
+            }
+            if (isset($wpfep_options['register_page'])) {
+                wp_delete_post($wpfep_options['register_page'], true);
+            }
+            if (isset($wpfep_options['edit_page'])) {
+                wp_delete_post($wpfep_options['edit_page'], true);
+            }
+            if (isset($wpfep_options['profile_page'])) {
+                wp_delete_post($wpfep_options['profile_page'], true);
+            }
+        }
 
         // Delete Options.
         delete_option('_wpfep_page_created');
